Rename incomethree to texitcoin.html
TEXITcoin is poised to lead the digital currency landscape, not just because its technology surpasses that of other top-tier proof-of-work cryptocurrencies, but because its success will directly and positively impact the Texas economy through spending and community trade. While Bitcoin serves as digital gold and Ethereum operates as a decentralized computing platform, TEXITcoin aims to achieve top rankings due to its practical usability in the marketplace.